BuilderPattern

+++ src/server/Channel.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Channel.cpp                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cdomet-d <cdomet-d@student.42.fr>          +#+  +:+       +#+        */
+/*   By: csweetin <csweetin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/26 14:31:43 by aljulien          #+#    #+#             */
-/*   Updated: 2025/03/19 14:19:39 by cdomet-d         ###   ########.fr       */
+/*   Updated: 2025/03/19 17:04:52 by csweetin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -82,7 +82,7 @@ std::string Channel::getTopic() const
 {
 	return (topic_);
 }
-int Channel::getMaxCli() const
+size_t Channel::getMaxCli() const
 {
 	return (maxCli_);
 }
